requirements.txt
fill_voids==2.0.5
fire==0.5.0
glymur==0.12.8
gpustat==1.1.1
GPUtil==1.4.0
h5py==3.9.0
matplotlib>=3.7.0
numpy>=1.22.4
pandas>=1.5.3
Pillow==10.0.0
sauth==1.1.0
scikit_learn>=1.2.2
scipy==1.11.2
setuptools==68.0.0
pathpretrain==0.1.2
ninja
detectron2 @ git+https://github.com/facebookresearch/detectron2.git